testmc.i8080: 8080 simulator
testmc.i8080.{machine,opcodes.opimpl} is copied from testmc.mc6800 and
modified appropriately. This is the core code for instruction dispatch, the
OPCODES map, and the Instructions map. Of all of this, only the OPCODES map
itself has any huge differences; significant parts of the rest look as if
they should be pulled up to testmc.generic.{machine,opcodes.opimpl}.

i8080.opcodes introduces something we didn't see in the
mc6800.opcodes: parametrized calls into the i8080.opimpl functions
that actually do the work. This grew mainly out of having more
registers on which we need to do the same function, e.g., _incr() on A
B, C, D, E, H, and L rather than just A and B. For the 8080 these
functions live in opcodes, rather than in opimpl, though it's not
clear in which one they would better live, or how to deal with the
naming of this.

This also introduces a separate file of tests, testmc/i8080/nativetest.pt,
for the tests that run the native code testers (see below). These creates
an environment that emulates enough of CP/M to make the programs work.

Under psrc/testmc/i8080/programs/ there are a number of programs (binary,
and source where available) that the above can use to to confirm parts of
the 8080 simulator behaviour. All should pass, but not all do, and some are
skipped because they can take a _very_ long time to run. These are
described in detail in i8080/programs/README.md.

Other issues include:
- The H flag on sub/cmp is definitely not working, and the H and C flags in
  general are not fully confirmed to be correct.
- programs/8080EX1.COM is not completing, no matter how long we let it run.
  This needs to be investigated.