WIP testmc.i8080.{machine,opcodes.opimpl} Code copied from testmc.mc6800
This is the core code for instruction dispatch, the OPCODES map, and the
Instructions map. Of all of this, only the OPCODES map itself has any real
differences; the rest looks like it should be pulled up to
testmc.generic.{machine,opcodes.opimpl} once we have a few instructions
implemented and it's passing some reasonable number of tests.

WARNING: No test classes have been added for any of this. However, the
parent repo src/e8_8080.pt is now producing an InvalidOpcode exception,
as it should be until we get $3E (ld A,d8) implemented:

    testmc.i8080.opimpl.InvalidOpcode: op=$3E,
	i8080 pc=0000 a=FD bc=0000 de=0000 hl=0000 sp=DFFE sz-h-p-c