README: Bring more into line with the current state, and contact info
This is far out of date, and still includes a lot of information from the
`8bitdev` repo that needs to be cleaned up. This minimal update should at
least fix the very worst of the problems with it, and a warning has been
added to the top of the file to make it clear that it's still well out of
date.