t8setup: Improve comment to suggest submodule initialisation
With a standard `git clone` of the parent project, the submodule will
not yet have been checked out, causing the attempt to run t8setup.bash
to fail. Add a little bit more code to the suggested parent ./Test
initialisation to init the submodule if it's not yet checked out,
which we check by looking for t8dev/t8setup.bash.