t8setup.bash: Improve modified Git submodule warning
The previous warning suggested runnning `git submodule update --init`
because it was assumed the most likely reason the developer was getting
this warning was that the t8dev submodule had not been checked out. That's
obviously no longer the case since it's code in the t8dev submodule that's
emitting the warning.

What the warning means now is that the t8dev submodule (or other
submodules) has been modified, which is generally because the developer is
modifying it himself, either to change the code or test a new update. In
that case we very much _don't_ want to re-init the submodule, as that could
lose the changes. So instead we just mention that submodule(s) are modified
and leave it at that.