t8dev.cli.t8dev: Non-"B8_HOME" discovery of t8dev etc. tests
Previously we were adding path.b8home('pylib') to the the default set of
dirs in which to do test discovery. This will no longer work when we remove
B8_HOME and path.b8home() after we start installing t8dev to the
virtualenv. Further, even if we did, this still misses other
submodules/packages such as r8format.

Ideally we should be finding all the Python distribution packages in our
working copy (t8dev, r8format, and anything else) that have a
pyproject.toml and adding (with proper munging of the relative location)
the test paths from tool.pytest.ini_options.testpaths entry.

This is a lot of work, however, so for the moment we simply look for any
pylib/ directories in our working copy and add those to the list of
dirs on which to do test discover. This is basically just what we were
doing before, except it now finds the r8format unit tests as well as
the t8dev unit tests.

(This still doesn't address the functional tests from the sub-packages,
which may need an entire re-think. Possibly we should be dropping the shell
script/runnable file way of doing that and porting those in some way to
pytest as well?)