pylib/pytest_pt.py: Removed, as it's supplied by r8format
This also fixes issues with running on newer versions of pytest and/or
Python, as r8format has the newer version of this module that supports
them.

This also re-emphasises that r8format is a hard requirement for t8dev,
because _nothing_ will run without the module above, whereas not all the
code and tests have dependencies on things from r8format.