t8dev.path test_addbin(): Copy the test dict because the function modifies it
One of the parameters to the parametrized test function is a dictionary;
this is passed to addbin() which modifies it. Under pytest 5 there was no
issue with this, but under pytest 8 this fails because somehow the test run
we see is getting a copy of the dict that's already been modified from a
previous test. (It's not clear how this happens; I don't see any reason the
test would be run twice.)

At any rate, it's good practice anyway to make a copy of the dict so that
we don't modify the parameter we've been given, which is what we do to fix
the problem.