testmc.generic.machine: Add half-wide register aliases to Machine
These are the aliases created by Reg('hl', width=16 split8=1) and the like.
The 'h' and 'l' aliases for the two halves of the 'hl' register were
created on the Registers() object, but not on the Machine() object (or the
object in a Machine holding the register attributes). Now they
are created.

Also:
- In the generic.machine tests, make the test Machine subclass __init__()
  call the superclass __init__(). We do it in the real code, so it must
  just have been forgotten in the test code.
- Extract a static method Registers.init_split8(cls, regspec) which is now
  used by both the Registers and Machine init methods to set
- Add a __repr__() for Reg so that we can see what it's been defined with
  when doing debug printouts.