pylib/ → psrc/: Rename Python code directory
This is now really a src/-style dir rather than a lib/-style dir since we
started putting the code in a virtualenv rather than running it directly
from the directory. And `psrc` is better for command line completion than
`pylib` because we now have other `py*` files in the top level of this repo.

This breaks existing .build/virtalenv installations; remove that dir in
order to fix this (or just do a full clean `./Test -C`).