t8setup: Do not exit on error, return instead to avoid killing caller
Since this script is always sourced, an `exit` will exit the _calling_
shell, which is obviously not what we want. Instead, we need to
`return` with the error code.

But we can't do this from a shell function like die() becuase that
will just return from the shell function and carry on (unless the
caller has -e set, which we can't rely on). Since we do this in only
one place anyway, just get rid of die() and explicitly return an error
in the one place where we need to do this.