t8setup: Add $T8_PROJDIR/bin/ to path only if that dir exists
We also move this to live with the $BUILDDIR/.tool/bin/ setup, since these
should pretty clearly be together. Though it's not entirely clear that we
should be doing this before we deal with the arguments and whatnot, rather
than after. But right now, nothing after this can fail anyway.