t8dev.toolset.setup: source_sha is now optional and we check HTTP status
The `source_sha` variable was previously required so that we could
verify that we'd downloaded the correct file. However, for some
projects (such as CSCP[1]) we're not too concerned with having a
particular version, and are happy with just downloading the latest
version.

Now that we don't require `source_sha`, however, a failed download may
give us, e.g., an HTML "404" page because we were never checking that
the HTTP request succeeded. We now check the HTTP result code and fail
if it's not a 200/OK.

[1]: http://takeda-toshiya.my.coocan.jp/