t8dev.toolset.cscp: Download/install Common Source Code Project archive
The Common Source Code Project is a large library of emulators for
various Japanese computers. We download the binary archive of
pre-compiled Windows programs because we can't build them on Linux
(they're Windows-only) but we can run them under Wine.

This requires an external archiver program; we add the PyPI `patool`
package which finds one for us. The user still has to install one of
the programs it knows about (see the header comment for details), and
the error messages are not great if it can't find one. This is all
really annoying, but the other options are worse (more on that too in
the header comment).

The Windows 10 binaries are unpacked to …/extract/cscp/ and then
copied to …/bin/cscp/. We install into a subdir of bin/ because there
are a lot of these and they're not runnable from the path anyway (at
least in Unix).

The binary release of CSCP includes Windows 10  and Windows Vista
binaries; we choose Windows 10  because, well, we don't know and that
sounds better.

Note that these are Win32 binaries: you will need the 32-bit version
of Wine to run them.