t8dev.toolset PREFIX_SUBDIRS: Add `extract/` for archive contents
We sometimes need to extract all the contents of an archive of executable
files even through we don't want to install all of them. This provides a
place to store those files from which we can then copy what we need.

This will be used by the CSCP archive download/install in the next commits.