t8dev.toolset.cscp: Install Windows 10  binaries to bin/cscp/
This requires an external archiver program, which is really annoying, but
the other options are worse. See the header comments for details.

We install into a subdir of bin/ because there are a lot of these
and they're not runnable from the path anyway (at least in Unix).

The binary release of CSCP includes Windows 10  and Windows Vista binaries;
we choose Windows 10  because, well, we don't know and that sounds better.

Note that these are Win32 binaries: you will need the 32-bit version
of Wine to run them.