t8setup: If T8_PROJDIR is not set, try to find it so we can set it
We look in each directory from the curent one upwards for a
`.build/virtualenv/` that appears to have t8dev installed in it. (We
currently check by looking for the t8dev executable in bin/ or Scripts/.)
If we find one, we use that directory. If we get to the top of the
filesystem without finding one, we just abort with an error as normal.

Also, fix a warning to go to stderr.