pyproject,t8setup: Switch to using Python packaging dependencies
This means that we no longer need users (or ourselves!) to specify the
t8dev package's dependencies in requirements.txt files.

Many of the dependencies are optional so that people wanting to use only
parts of t8dev (e.g., just the testmc simulators in this repo) need not
bring in the other dependencies. (See pyproject.toml for the details of
what dependency subsets are available.) However, most users will probably
want all the standard dependencies (which anyway are not very large), so
t8setup.bash now installs `t8dev[all]` to bring them all in.