psrc/pytest_pt.py: Restore pytest_pt because of r8format issues with it
For some reason I cannot convince the Python packaging `build` program
to include top-level modules that are files, rather than directories,
in the binary distribution of r8format.[1] This means that r8format
does not make the pytest_pt module available unless installed from a
directory or network URL using the -e (editable) option.

(This wasn't noticed until now because nobody was doing done non-editable
installs, and it's only been about a day since r8format became available as
a binary package.)

To work around this for the moment I'm restoring the second copy of it
in this repo. Hopefully we can later work out just what the problem is
and fix it.

Note that we have a vaguely similar issue with pactivate (which I also
haven't figured out how to include in the r8format binary
distribution), but that one has an insoluble circular loop problem: we
need pactivate to create the virtual environment into which we want to
install r8format to provide pactivate. So there's really no way of
getting around having our own copy of that.

[1]: https://stackoverflow.com/q/78376273/107294