WIP cli.t8dev.emulator: copy CSCP binary to emulator run dir
CSCP emulators look for their ROMs in the same directory as the program
itself (*not* the CWD!), so we make a copy of it in the directory we set up
to hold the ROMs/files/etc. for this emulation run.

We copy instead of symlink because Wine, emulating Windows *really* well,
hates symlinks.