t8dev.cli.t8dev.emulator: Error messages/exits when file/URL not found
This now catches errors in loading/downloading both the initial RomImage
loadspec and (as with the previous commit) any patchspecs applied to it
after, and exits with an error message.

We do not give the patchspec ROM name that generated an error because
we don't know it (we do all patchspecs in one call and it's not in
the raised exception), but this doesn't matter much since we do give
the URL or file path that failed, and the user can easily identify the
bad argument from that.