t8dev.toolset.asl: Now build `upstream` branch of `asl-releases`
Previously we built the `dev/cjs/current` branch, which I was updating
from time to time. But after the spate of various user interface
improvements over the last few years releases have slowed and are now
focused more on adding CPUs and the like, so it seems reasonable
always now to use the latest release (or at least the latest available
in `asl-releases`)..

We should probably at some point add a parameter to `t8dev
buildtoolset asl` that would let us select which release we want.