psrc/pytest_pt: Replaced by a runtime dependency on `pytest_pt` package
Since pytest_pt is now available from PyPI as a distribution package,
we can now use that rather than including a duplicate copy of the
pytest_pt module here.

This is a runtime dependency rather than a development dependency for
reasons explained in the pyproject.toml. This also maintains backwards
compatibility for users of this module who were depending on it to
make the `pytest_pt` module available.

Version bumped to 0.2.0.dev1.