cli.t8dev.emulator: New locations for TK-80BS ROM images
The retroabandon/tk80-re repo was re-organised recently, particularly
to indicate that the "TK-80BS" ROM images are not the original images
but a clean-room re-implementation, in other words, a clone. So we
now use the clone images by default until we can get around to dumping
the original images.