testmc/*/tmc/bios*: Move tmc BIOS code from 8bitdev to t8dev
The source code for the "tmc" unit test and command line simulator
BIOSes was previously in cjs's personal repo, 8bitdev; it was not kept
with t8dev when this project was extracted because at the time we had
no way to build it from t8dev.

These now are moved into the Python code directories as "resource
files"; the paths to them can be found using importlib.resources.
(This requires that they exist in the filesystem, which they do for
standard and editable package installs. This will not work for other
storage methods for package source code such as ZIP, but we have no
intention of doing that.)

This forces our minimum Python version to 3.9 (it's not worth writing
compatibility code to emulate importlib.resources to support earlier
versions of Python), and we add setuptools-scm which handles ensuring
that the non-Python files are included in the package. (This includes
the unit test .pt files as well, but that's fine; developers should
actually have the option to run them even from standard installs.)

To build these, we add a new `t8dev` subcommand, `aslt8src`, which
finds all *.asm files under the t8dev packages (currently `t8dev` and
`testmc`) and builds them into the main project's .build/obj/
directory. (This must be run before `tmc` so that the object files are
available to be loaded by it.)

We also update some paths to use this:
- The ASL build function adds the directory above those two
  package directories to the ASL include path so you now
  `include "testmc/mc6800/tmc/bios.a68"` or similar.
- The pytest loadbios() fixture loads from the new paths.
- The `tmc6800` program uses the new paths. (This will so be
  replaced by a more generic program that allows selection of
  the CPU simulator.)