testmc.tmc: New command-line simulator runner to replace tmc6800
This new `tmc` (testmc.tmc) program is a more generic replacement
for tmc6800, able to use any of the CPU simulators.

It takes a CPU string and a path. The CPU string is matched as a
substring against the list of simulators, using the first one that the
substring matches. The path processing follows similar rules to tmc6800,
adding a `.p` if there's no extension and, if there are no slashes in
the name, looking under .build/exe/tmc/CPUNAME/ for the file.