testmc/i8080/tmc/: i8080 unit test and `tmc` BIOS
This is basically just a straight port of the 6502/6800 BIOSes, with the
following changes:
- The current address used by the `assertdef` macro is `$`, not `*`. (Using
  `*` doesn't generate an error, but silently ignores any code overlaps!)
- The BIOS is moved to $80, above all the 8080, 8085 and Z80 interrupt
  entry points.
- The character I/O port remains memory mapped because we don't yet
  support the I/O address space, and is at memory address $FF.

The unit test fixture and tmc both automaticlaly know the memory
address used for character I/O because they read the `charinport`
value from the symbol table of the program under test. (It's assumed
that that program has included `bios.i80` or whatever is appropriate
for the CPU which defines that.)

Unfortunately this does _not_ implement any way to exit the i8080
command-line simulator because we can't use the "execute illegal
instruction" mechanism to generate an exception, the 8085 not having
any opcodes that are illegal instructions. For the moment you must
terminate it using the Unix `kill` command. This will be fixed in a
later commit.