testmc.tmc,bios.i80: Implement exit port for i8080, and all CPUs
The 8085 (which the 8080 simulator will soon simulate) has no invalid
instructions, so the trick of executing an invalid instruction to
cause the simulator to throw an exception and exit doesn't work.

Instead, we define an I/O port where the "virtual device" exits the
simulator if a specific value is written to it. We change the the 6800
and 6502 simulators to work the same way since this is more convenient
than dealing with different ways of exiting for different simulators.