testmc.generic.machine.pt: Rename test machine classes
Rename the local `m` fixture here to `TM` ("Test Machine class"). Unlike
the standard `m` fixture in testmc.pytest.fixtures, this is not a Machine
instance but a Machine subclass. Using `m` for this was confusing,
especially since we re-assigned an instance of it to the local `m` var in
most tests.

Also rename a Machine subclass in one of the tests for similar reasons.