testmc.generic.machine: Add biosname()
This returns the name of the `tmc` BIOS to load. It defaults to the
(non-qualified) name of the package in which the `machine` module
(containing the Machine class) resides, but can be overridden should
two different Machine subclasses want to share a BIOS.

This will be used to avoid having to specify the machine name every
time for the pytest.fixtures.loadbios() fixture.