fixtures.loadbios(): New API making `biosname` parameter optional
This is incompatible with the previous API; so we bump the 0.x version.

This also requires that projects using this add `t8dev/psrc/` to their
`t8dev aslauto` parameters. (Possibly this could be done automatically,
but we'd need to figure out how to determine if we're using t8dev source
via a submodule, and how to find the path to it if we are.)

We no longer need to specify the `biosname` parameter to the
testmc.pytest.fixtures.loadbios() fixture because it can be determined
automatically. (You can still specifiy it as a named parameter if you have
need to use a different one.) But becuase this was the first positional
parameter, and the second positional parameter was the input bytes, to
avoid inadvertantly assigning one string to the other we now disallow
positional parameters entirely and you must specify `input=b'…'` for that.

We also weren't testing this fixture, but that's easy to add since we
already had tests (that weren't using the fixture) for all three BIOSes
(testmc/*/tmc/bioscode.pt); we just change those to use the fixture rather
than doing the same thing by hand. These tests are almost entirely
duplicate code, and we should extract the common code for those.