setup.cfg: Install non-Python scripts from bin/
Previously we were not including these in the distribution package
because `setuptools` has no way to specify this in the
`pyproject.toml`. These scripts are, however, potentially useful, and
we have further definitely useful scripts that we're going to add.

setuptools creates a build directory when you install from source;
The `build-base` build option changes the default directory from
build/ in the root of the project dir.