mos65/machine: Delay py65 ModuleNotFoundError until it's actually needed
The testmc changes introduced in 0.0.5 include the module having a list
of all the simulators, which causes all simulator modules to be loaded.
That caused errors when this package was used in a configuration without
py65 available because the mere load would throw a ModuleNotFoundError
for py65. Now we delay that error until someone actually tries to _use_
the mos65 simulator, so that py65 remains unrequired for those not using
that simulator.

We really should have some tests that test the install configurations
other than t8dev[all], or at least ensure that every use of a module
module in the "optional dependencies" list is lazily loaded like this.