testmc: Change call() to execute return; fix i8080 machine bug
This is unlikely to affect most tests, but some can be, so this
bumps the version number to 0.2.x.

To determine when we're returning back to the unit test framework we
previously stopped on any return opcode given in the _RTS_opcodes
list, and checked to see if that would return to the `retaddr`
parameter given to call(). However, this doesn't work with conditional
returns (such as the i8080 has) since if we stop before executing the
conditional return we don't know if it will return or not.

We solve this by actually executing all RET instructions and instead
of stopping on RET instructions, stopping at the `retaddr` location
and checking to see if the stack is "empty" (i.e., the SP is pointing
to the location it was when call() started).

This has the disadvantage that we can no longer check to see if we
have an "unexpected stack" on return, such as when the return address
placed on the stack by call() was changed. These sorts of situations
would be a bit pathological anyway, so it's probably not a big deal.

This commit also fixes an off-by-one error in i8080.machine's
pushretaddr() and getretaddr(). These seem to have gone undiscovered
because it didn't matter until the new version of Machine.call()
wanted to use them in a different way.