API! testmc/*/tmc/bios.* → biosdef.*: Rename definitions-only files for clarity
When `include`ing files it's important to know whether the file
generates any code, or it's just symbol/macro/etc. definitions. In
languages like C this is done by using a .h instead of a .c extension,
but we also need to know what CPU the defs file targets since some
definitions may be machine-specific. We examined several ideas for a
naming convention, such as .hi80, .h.i80, .i80.h, and so on, but
conventions like this seemed a bit awkward, so we settled on appending
`def` to the basename (before the extension) of the file.

This is a breaking API change, but the 0.x version number bump is
piggybacked on the previous commit, as that already bumps it.