testmc.generic.iomem: Throw exception when an I/O function returns a bad value
When responding to a read for the memory location to which it's been
attached, an I/O function _must_ return a valid byte (or throw an
exception). If it doesn't do so, the invalid data will propagate who knows
where in the machine and probably generate a very mysterious exception
later.

On reads we now check that the returned value is an int in the range $00 to
$FF, and throw a TypeError (if the value is not an int) or ValueError
(if the value is an out-of-range int).