testmc.tmc: Make exitport readable
This was causing the command-line simulator to crash when reading that
memory location because it was returning None instead of a byte. We now
return the exit command byte in case you want to see what it is, and to
make the port easier to identify in memory dumps.

We still have the issue where if you try to examine location $FF it
waits there for you to type a key, because that's the "read the
keyboard" location. This will be fixed later when we move to using
I/O address space for these, instead of memory address space.