pyproject: Make all optional requirements standard
Previously you had to remember to install `t8dev[all]` if you wanted to use
all facilities of t8dev, or remember which of the various optional
subgroups you needed. Since none of the optional requirements were terribly
large (most are almost trivially small), it's not worth trying to save a
trivial amount of time on installation and so we just remove all options
and install everything.

Installing `t8dev[all]` still works, but produces a warning that the
package "does not provide the extra 'all'".