t8setup: Do not add `[all]` to t8dev package; no longer needed
It's no longer needed nor supported. And we remove the extra code that
allows us to have an "extras" option appended to the package name because
this was the only thing that was using it. (Anybody who still needs it for
their own stuff would probably be putting the install line into
requirements.txt, anyway, which possibly we should be doing.)