t8setup.bash: No longer install Python packages in subdirs
You now need to add them explicitly to `requirements.txt` in the parent
project with lines like `-e pkgdir/`.

This code was complex and not terribly reliable. Using `requirements.txt`
is not only simpler and always works properly, but it also consolidates the
project's package dependency list in one place.