t8setup.bash: Do not remove the -- from $@ if we encounter it
After we've finished with all our options (just -C and -c), we leave
the remaining options in $@ so that the caller can use them.
Typically, the caller will want to pass these on to pytest.

We immediately stop parsing options if we encounter a --, so that -C
and -C can be passed on to something else if not meant for us. But if
the caller also needs to separate additional `Test` options from those
passed on to e.g. pytest, it needs to know that a -- was encountered,
so we now leave that at the front of $@ instead of removing it.

Callers that have no options of their own can safely just pass $@ to
pytest, as it seems to ignore any -- options and continues to parse
options after that.