cli.t8dev.emulator: RunCPM does not need symlink in CWD; doc updates
We do not need to link the RunCPM binary into `emudir`; we can run it
from anywhere because for its configuration it cares only about what
the CWD is, not where the binary is.

We still have a separate way of setting `emudir` because, unlike CSCP,
this suite *is* the emulator, rather than having emulators underneath
it, and that is reflected in the .build/emulator/ directory structure.
It seems to make sense to pull up an emudir() into the superclass, but
we need to figure out how to handle the "has/has not multiple
emulators in the suite" thing, and of course write some tests for
how we handle this.

Here we also slightly improve the comments in the file.