cli.t8dev.emulator.RunCPM: Copy files to to RunCPM "disk"
This copies any files given on the command line to drive A0:, and, if
available, the osimg:cpm/2.2/ files to drive C:, so that DDT, STAT
etc. are available if you want them.

Experimentation showed that RunCPM doesn't deal well with files that
have lower-case names: DIR will list `foo.com` as `FOO.COM`, but
typing `foo` or `FOO` just produces `FOO?`. So we convert all filenames
to upper case, without checks to see if this might overwrite files
that differ only in case.

We considered using symlinks back to the files rather than making a
copy, but it seems better to have a separate copy so that if the
program modifies any of the files on the disk we can compare those
with the original versions.

Since we now have a reasonable amount of functionality here, update
the changelog and bump the dev version number. (We may add more
functionality before actually doing a release.)