cli.t8dev.emulator.RunCPM: Disable clear screen when RunCPM starts
RunCPM, for no good reason I can see, clears the screen on startup on
POSIX systems. This is annoying because it can erase useful output
from the build system. It uses termios for this, so we disable it by
setting TERM=dumb in the environment.

This required extending t8dev.tool.run.run() to accept an `envupdate`
parameter with the keys to update and their new values. We need to do
a bit of work to do this update (with the new newenv() function)
because subprocess.run() must be given a complete environment if it's
specified, and it's best that all other environment variables be
inherited in case they're being used by anything. (RunCPM needs PATH,
for example, because it does a system('clear').)