cli.t8dev.emulator.RunCPM: Add =a option to auto-run file and exit
With this option set we run the first file given on the command line
by creating a $$$.SUB file on the root of A: drive containing (properly
encoded) FOO.COM and EXIT commands, just as SUBMIT.COM would do.

We need to use = as our option character for the moment because all
options starting with - are consumed by the top-level t8dev options
parser. This should be fixed at some point by creating sub-parsers
for the various subcommands.