testmc/i8080/bios/cpmcode: prchar/rdchar/etc. implmentation over CP/M
Output seems to work fine, but this doesn't handle input correctly because
it's using CONIN which echos the read character (as well as apparently
swallowing characters with a special meaning to CP/M, such as ^S and ^P).
An upcoming commit will fix this by using direct console I/O for the read.

The new bios/ subdir under each CPU will contain the appropriate files
for use of BIOSes on other systems: SYSNAMEdef.asm where we need only
definitions, and SYSNAMEcode.* where we actually need code to transform
the API calls, as we do here.