testmc/i8080/bios/cpmcode: Use direct console I/O for reads
This is working better now, but there are still issues with printing
backspace in 0cjs/8bitdev's tmon where a backspace is not not being
printed here where it is under the tmc emulator. It's not clear
whether the problem is in tmon or this BIOS; some basic tests of the
BIOS don't show the problem.