testmc/*/bios*: Make A destroyed, both in API docs and reality for tmc BIOS
We've recently been seeing bugs in code that assumed that A was preserved
by `prchar` and the like, which it happened to be by the tmc BIOSes, but
often is not by real-world BIOSes. Clarify the API docs to indicate that
A may be destroyed, and change the tmc BIOS implementations actually to
destroy it so that we see these bugs sooner rather than later.