API cli.t8dev/: Rename subcommands and use argparse for all of them
This is a very large change and it exposes somewhat our need for
better testing of t8dev subcommands. It does however work for
everything that github.com/0cjs/8bitdev uses (which is most of it) and
has seen some manual testing as well.

This replaces all ad-hoc parsing done by subcommands with ArgumentParser
subparsers, which now gives us subcommand-specific options (such as `t8dev
emulator runcpm --autorun`), better help, and so on.

There is an issue with subcommands where if one is missing a poor error
message is produced: 'the following arguments are required:', without
saying which arguments they are. This has been with something of a hack at
the top level, but not for subcommands of subcommands. Probably a common
function to deal with this should be extracted.

This has introduced enough change that we now need a fair amount of
refactoring to bring the code back to a nicer state, such as removing
t8dev.cli.t8dev.shared.ARGS which are no longer needed except for the one
verbosity setting of the vprint() function.