cli.t8dev.emulator.*: Extract modules for individual suites
..emulator.suite contains the Suite superclass, and CSCP and RunCPM
have been extracted into separate modules. The remainder of the
suites, which are incomplete and don't work, have been dropped into
the ..emulator.other module for the moment. We'll move those out into
their own modules when and if they grow significantly larger.

The Suite superclass is wanting to have common code from CSCP and
RunCPM moved up into it.