cli.t8dev.emulator.suite: Pull up emulator and emudir()
self.emulator is now automatically set based on the parsed command line
arguments (to args.emulator, or None if that doesn't exist).

self.emudir() creates the emulator directory `.home/{suitename()}/`,
or `…/{emulator}/` under that if emulator is set, and returns that
and any further components specified as arguments.

Clean up some code related to this in emulator.cscp, and also remove
incorrect docstring for setup_emudir() and add one for set_bindir().