t8dev,testmc: Tweak a few things to work with Python 3.9
t8dev.cli.t8dev.emulator.virtualt cannot use contextlib.chdir, as
that's not available until Python 3.11. Replacing this with a couple
of os.chdir() wrapped in a try/finally works just as well, if more
verbosely. (If we need to do more of this, we should probably build
our own decorator for it. StackOverflow has several examples.)

testmc.tmc was using a `… | None` in a type description, which is not
avilable until Python 3.10. Mypy provides an `Optional[…]` constructor
that does the same thing, and in some ways is more clear.

Tested on Python 3.9.18, 3.10.13 and 3.11.2.