t8dev.toolset.asl: Use `main` branch Makefile.def
We were previously copying the sample file for x86_64-linux,
`Makefile.def-samples/Makefile.def-x86_64-unknown-linux`. This doesn't
work on ARM Linux systems, unsurprisingly, and we really want to be
using our custom Makefile.def from the `main` branch of `asl-releases`
anyway. So we now use Git to pull that out of the `main` branch and
write it to our working copy, first prefixing it with an __INSTDIR
setting for our t8dev tools install location.