testmc.tmc: Raise proper EOFError() when reading at EOF
The BIOS `rdchar` routine has no facility to signal the end of input
since on a real machine the routine would simply block until someone
types something (or turns the computer off).

But we do need to handle this for cases where someone's redirected
stdin from a source that can reach EOF, so we now check to see if we
actually got a char and, if not, we raise an EOFError() that exits the
simulator.