* → t8dev/: Move package to subdir
This moves only files specific to the t8dev package (though that's
most of them), leaving at the top level files that will be common when
we bring the r8format package into this repository.

This brings in a slight annoyance of having to go through yet another
directory when we're editing source, but in the long run we intend to
fix that with a top-level src/ directory containing symlinks to the
Python import packages under */psrc/. And that will also get rid of
the completion annoyance with having `t8dev.egg-info/` right beside
`t8dev/`. (But we need to make sure that Windows users don't use the
copies that will appear there instead of symlinks.)

The [pypi-release] `build-release` script has been updated so we can
now release packages in subdirs: `…/build-release t8dev`.

[pypi-release]: https://github.com/cynic-net/pypi-release.git