r8format/Functest: Only functional tests now; Test calls this
The top-level Test has always been running all the unit tests in this repo,
including the t8dev and r8format subdirs. We now rely on that.

r8format/Test is moved to r8format/Functest and its unit-testing
functionality is removed. The top-level Test script calls it when the unit
tests are complete and pass.

r8format/Functest still calls r8format/psrc/cmtconv/Test, which does the
functional tests for cmtconv. That now assumes (with a check) that
$T8_PROJDIR is set (so it knows where to put files generated by the tests)
and that the virtualenv with cmtconv is already set up, rather than setting
up its own.

r8format/requirements.txt is removed because now the only virtual
environment we use is at the root, it already brings in `pytest` and
`pytest_pt` as package dependencies, and `build` and `twine` are now
brought in independently by [cynic-net/pypi-release] when we need
to do a release.