src/: Remove to avoid pytest collection problems
Depending on the configuration, pytest may try to collect files under
the src/ directory, following the symlinks and not only running tests
twice unnecessarily, but also for the `testmc` package producing,
"ERROR t8dev/t8dev/psrc/testmc - ValueError: Plugin already registered
under a different name: …". (It's not clear what configuration change
triggers this; it happens in 0cjs/8bitdev, but not in some other users
of t8dev.)

So we remove the src/ directory and its symlinks. Those who want the
symlinks, and whose configuration supports it, can create them from a
directory in their own project dir.

This does not fix the problem; this still breaks for e.g. 8bitdev even
with the symlinks in a project directory subdir. It's not clear yet
what we can do to fix that. Possibly it would make more sense just to
use a command-line tool that knows how to find source code files deep
in subdirs without having to type/complete all the intermediate
directory names.