Test pytest: This script always uses this dir as rootdir with no pytest config file
See the comments in Test for details of why we do this.

Now that we look at this more closely, it's not clear if we should
actually be supporting using a different, externally set T8_PROJDIR in
this script. That was originally because we needed our r8format
dependency installed and might want to use a different one from the
release version, but now that both packages are in the same repo, we
should probably always be using the r8format from this repo.